requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
qrcode==8.0
Pillow==11.0.0
requests==2.32.3
reportlab==4.2.5
gunicorn==23.0.0
xendit-python==7.0.0
python-dotenv==1.0.1
